requirements.txt
antlr4-python3-runtime==4.13.2
numpy==2.2.5
pytest==8.3.5